First commited as before_10017_rev2

+++ org.eclipse.jgit/src/org/eclipse/jgit/storage/file/PackBitmapIndexBuilder.java
@@ -215,7 +215,7 @@ public class PackBitmapIndexBuilder extends BasePackBitmapIndex {
 	}
 
 	/** @return the index storage options. */
-	public byte getOptions() {
+	public int getOptions() {
 		return PackBitmapIndexV1.OPT_FULL;
 	}
